Fix P2 findings: UltimateInterface auth races, DataTransit streams, DataProtection cast
UltimateInterface (P2-3299, P2-3300):
- TableauStrategy: add SemaphoreSlim _authLock (1,1) to serialize
  EnsureAuthenticatedAsync — concurrent callers raced on _authToken/_siteId
  with check-then-act, potentially overwriting each other's site ID on sign-in
- MetabaseStrategy: add SemaphoreSlim _authLock (1,1) with double-checked
  locking pattern in EnsureAuthenticatedAsync to serialize session token fetch
- ApacheSupersetStrategy (OpenSourceStrategies.cs): same pattern for
  _accessToken TOCTOU race during concurrent Superset login

UltimateDataTransit (P2-2666):
- CompressionInTransitLayer.TransferAsync: wrap compressedStream in
  try/finally so DisposeAsync is called even when _inner.TransferAsync throws
  — previously the MemoryStream/GZipStream leaked on any transfer error
- GrpcStreamingTransitStrategy, Http3TransitStrategy: minor delta fixes
  from prior stash (caught by build run before commit)

UltimateDataProtection:
- VersioningSubsystem: explicit (int) cast on versionNumber to resolve
  implicit narrowing conversion warning in CreateVersion

+++ Plugins/DataWarehouse.Plugins.UltimateDataProtection/Subsystems/VersioningSubsystem.cs
@@ -52,7 +52,7 @@ namespace DataWarehouse.Plugins.UltimateDataProtection.Subsystems
             {
                 VersionId = versionId,
                 ItemId = itemId,
-                VersionNumber = versionNumber,
+                VersionNumber = (int)versionNumber,
                 CreatedAt = DateTimeOffset.UtcNow,
                 Metadata = metadata,
                 SizeBytes = 0,
